Extract the counting measurer into a shared test helper
