+++ ComposeTextEditor/src/desktopTest/kotlin/markdown/ImportRelayoutCostTest.kt
@@ -1,15 +1,12 @@
 package markdown
 
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.TextLayoutResult
-import androidx.compose.ui.text.TextMeasurer
 import com.darkrockstudios.texteditor.markdown.MarkdownExtension
 import com.darkrockstudios.texteditor.state.TextEditorState
-import io.mockk.every
-import io.mockk.mockk
 import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
+import utils.MeasureCounter
+import utils.countingMeasurer
+import utils.editorWithCounter
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -30,39 +27,6 @@ class ImportRelayoutCostTest {
 
 	/** A document with nothing to decorate skips the second pass. */
 	private val plainImportPasses = 1
-
-	private class MeasureCounter {
-		var calls = 0
-	}
-
-	/**
-	 * A [TextMeasurer] that tallies calls and hands back a one-visual-line layout, so
-	 * `updateBookKeeping` builds a full set of line offsets from what it returns.
-	 */
-	private fun countingMeasurer(counter: MeasureCounter): TextMeasurer {
-		val layout = mockk<TextLayoutResult>(relaxed = true)
-		every { layout.multiParagraph.lineCount } returns 1
-		return mockk(relaxed = true) {
-			every {
-				measure(
-					any<AnnotatedString>(), any(), any(), any(), any(), any(),
-					any(), any(), any(), any(), any(),
-				)
-			} answers {
-				counter.calls++
-				layout
-			}
-		}
-	}
-
-	private fun TestScope.editorWithCounter(counter: MeasureCounter): TextEditorState {
-		val state = TextEditorState(scope = this, measurer = countingMeasurer(counter))
-		// Book-keeping is suppressed until the viewport has a real size, so lay the
-		// empty document out first and count only what the import itself costs.
-		state.onViewportSizeChange(Size(800f, 600f))
-		counter.calls = 0
-		return state
-	}
 
 	/** Returns how many times importing [markdown] measured a line. */
 	private fun TestScope.measuresToImport(markdown: String): Int {
